Warn if using incompatible architecutre
Some packages can not be build on specific architecutres, e.g. because
Golang is not available. Until this commit the packages would just fail.
Now a adopted version of the CicleCI script is added which evaluates
first if a package can be compiled for the selected architecutre. If not
a warning annotation is added.

Signed-off-by: Paul Spooren <mail@aparcar.org>