Add CI
So that the results of this PR (and future PRs) can be seen directly. I
think more tests can be done, e.g. for the various build options, but I
didn't want to make this more complicated that it needs to be for now.
(I also chose to test with 3 architectures - in theory only one is
necessary, let me know if I should add/remove any.)

Signed-off-by: Jeffery To <jeffery.to@gmail.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [include commit message]